mci: core: don't fail MMC probe if HS200 isn't supported in HW
A MMC may support HS200 in theory, but it's unusable due to I/O
voltage. Unlike Linux, barebox doesn't check the I/O voltage, so it
will attempt enabling HS200 and get told no by the card.

This no currently triggered a failure of the probe:

  mmc1: detected MMC card version 5.0
  ERROR: mmc1: mmc_select_hs200 failed, error -5
  WARNING: mmc1: Card's startup fails with -5
  Cannot set parameter mmc1.probe: I/O error

But the correct behavior should be attempting the next speed mode.
-EBADMSG is the error code used to indicate this, so return that instead
when switching to EXT_CSD_TIMING_HS200 fails, because the card couldn't.

This is returned directly as to skip the error message in the err case,
which would just be confusing.

Fixes: ffa59e16bb96 ("mci: mci-core: add HS200 support")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240415052734.366242-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>