soc: rockchip: io-domain: handle missing supply correctly
The original Linux code calls regulator_get_optional and handles its
-ENODEV return code, which morphed into regulator_get when ported to
barebox. The error handling stayed as-is leading to causing supplies to
trigger errors instead of being silently ignored.

As we use NULL to describe the dummy regulator in barebox, we can add a
trivial regulator_get_optional implementation and use it to fix the
I/O domain driver.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240415053130.368278-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>