gpio: rockchip: fix support for RK3588/RK3566
The GPIO controller on newer Rockchip SoCs isn't compatible with the
older ones. Back when GPIOv2 support was added, RK3568 was the only
SoC supported using it, but since then support for the very similar
RK3566 as well as the newer RK3588 has followed, but the driver wasn't
extended to support them.

As GPIOv2 controllers have an identification register, start checking it
like Linux does and use that to detect whether we have a GPIOv2 or GPIOv1.

This issue was found trying to read GPIO inputs of the RK3566.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240415053137.368391-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>