ARM: dts: rk356x: describe serial reboot mode
The PMUGRF (Power Management Unit - General Register File) of the RK3568
has a general purpose register checked by the BootROM on power-on to
decide on whether to drop to recovery mode (rk-usb-loader/rkdeveloptool).

Describe this in the device tree, so it's possible to use, e.g.

  global.system.reboot_mode.next=serial reset

to force falling into serial (USB) boot mode.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240415053253.369354-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>