mtd: nand: mxc_nand: clear NAND_BBT_CREATE flag
On i.MX NAND the factory bad block markers are not at the place the NAND
core expects them to be, so the core also can't properly generate a bad
block table. Clear the NAND_BBT_CREATE flag to prevent the core from
doing so. This preserves the behaviour we had previously with the
barebox driver which was lost when we replaced it with the Kernel
driver.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>