ddr: imx: hide config prompt unless COMPILE_TEST
The DDR drivers are called into by low-level code for DRAM init.
The Kconfig options for them are selected by the SOC symbols, e.g.
ARCH_IMX8M and there's no point in asking for user decision: either
it's required and already selected or it will be unused and subject
to linker garbage collection. Therefore, hide the prompts by default.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240423060046.1513717-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>