mci: core: make execute_tuning mandatory for HS200
We assume that any host controller that doesn't implement execute_tuning
will have no issue with HS200 operation. This is likely untrue for most
host drivers that we have and it will confuse users:

  - Enable CONFIG_MCI_TUNING, e.g. for i.MX
  - devinfo shows HS200 operation
  - Actually reading blocks will fail with EIO

Therefore, let's make execute_tuning mandatory for HS200. If some driver
doesn't need it, it can always define an empty stub for it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240423062206.1518464-3-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>