ARM: i.MX8MP: fix compatible string for koenigbauer,alphajet board
During the review of the initial board support patch, the SoM compatible
was changed (and unified) to "congatec,qmx8p" from several different
incompatible variants. However, the compatible in the alphajet device
tree contained a different typo and was overlooked, which had the effect
that the barebox_update handler and the Ethernet fixups were not
registered on the alphajet board.

Fixes: 34f14b1519715d47d879 (2024-03-11, "ARM: i.MX8MP: add König+Bauer AlphaJet board")
Rubberducked-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240430102357.1357983-1-rhi@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>