kbuild: make implicit int to pointer conversion warning an error
We are guilty of using ints as pointers in readl/writel, but the old
definition of these uses a cast to silence the warning.

Everywhere else there's no reason to allow this, so make this a hard
error always to weed out any instances we may still have.

Disallowing this has the added benefit of making it easier to change API
in barebox: We do not guarantee out-of-tree board code API stability,
but we want changes in API to result in build errors, not in warnings
that may be eaten up by a build system sitting in-between Kbuild and the
user.

While at it, we fix the whitespace for the surrounding lines.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240430131540.3597376-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>