pblimage: ls1028a: fix handling of short reads on
Newer toolchains rightly complain:

  scripts/pblimage.c:259:17: error: ignoring return value of ‘read’ declared
       with attribute ‘warn_unused_result’ [-Werror=unused-result]
    259 |                 read(in_fd, mem_buf + 0x1000, image_size);

In the ls1046a case, this doesn't happen, because pbl_fget is used,
which does a number of things:

  - report warnings and exit on read errors
  - fill buffer with 0xff on short reads
  - increment pmem_buf and pbl_size by the number of bytes emitted

All of which, sound like a good idea for the ls1028a as well. Therefore,
initialize pmem_buf and pbl_size to the correct value and use pbl_fget.

This makes the code easier to follow through by making it explicit that
the image's payload is always given a 4K alignment, independent of the
size of the image header.

This change has only been build-tested.

Fixes: d5e5a65c2b64 ("pblimage: Add LS1028a support")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240515060758.3580734-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>