ARM: fix overlap between ramoops area and early malloc area
ramoops is usually a few megabytes and the early malloc area is just
128K, so the effect of this overlap was likely limited, but depending on
size of the data stored to pstore, this could still lead to problems.

Avoid this by locating the early malloc area before the barebox binary.
That way that space is reclaimed in barebox proper for the non-early
malloc region.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240517061057.3299857-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>