common: hide DEFAULT_COMPRESSION menu
CONFIG_DEFAULT_COMPRESSION_* refers to the compression done for
in-barebox binaries and is not useful when barebox proper as a whole is
compressed as in that case, we would have compression applied twice.

For this reason, DEFAULT_COMPRESSION_NONE is the only possible choice in
a PBL-enabled build. Having the choice with only one option is confusing
though, so let's hide the prompt.

Preferably, we'd omit these options in PBL builds completely, but this
seems not easily possible.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20240517062151.3301660-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>